bug fixed -Wextra

+++ p2/munch1.c
@@ -31,7 +31,7 @@ void* func_munch1(void* args)
     if(str[0] == '\0'){
       break;
     }
-    for(int i = 0; i < strlen(str); i++){
+    for(int i = 0; i < (int)strlen(str); i++){
       if(str[i] == ' '){
 	str[i] = '*';
       }
