+++ p2/munch2.c
@@ -28,7 +28,7 @@ void* func_munch2(void* args)
     if(str[0] == '\0'){
       break;
     }
-    for(int i = 0; i < strlen(str); i++){
+    for(int i = 0; i < (int)strlen(str); i++){
       if(islower(str[i])){
 	str[i] = toupper(str[i]);
       }
